% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBCarray.R
\name{SimBCarray}
\alias{SimBCarray}
\title{SimBCarray.R Simulate Array Data}
\usage{
SimBCarray(SimBCdes)
}
\arguments{
\item{SimBCdes}{A SimBCdes class object which contains:\cr
\itemize{
\item{N:} number of samples\cr
\item{Pblock:} number of feature blocks\cr
\item{BlockCovDF:} Pblock length DF\cr
 \itemize{
  \item{p -} number of features in block\cr
  \item{rho -} correlation\cr
  \item{scale -} scaling factor for covariance matrix\cr
  }
\item{DesignMat:} N x Q  design matrix [Non-Full Rank for convenience]\cr
  N samples and Q covariate design columns\cr
\item{BlockBetaMat:} Q x Pblock  matrix of model coefficients\cr
    valid per block.\cr
\item{DesignDF:} design dataframe with Q rows and columns:\cr
  \itemize{
     \item{name -} variable name\cr
     \item{type -} factor of numeric\cr
     }
\item{SampleNames:} N length character vector for the sample names\cr
\item{FeatNames:} P length character vector for the feature names\cr
}}
}
\value{
A list containing:\cr
        \itemize{
          \item{X:} P x N data matrix
          \item{PhenoDF:} N x Qunq data frame, where Qunq is the number\cr
             of unique variable names in DesignDF
              }
}
\description{
The functions simulates array data under a block diagonal
covariance matrix and a simple linear model for the sample
specific means.
}
\details{
add some details here
}

